removed editing subtask from task and replaced SubTaskMenu.kt with new compose
