+++ app/src/main/java/com/example/tidy/viewModels/AddTaskScreenViewModel.kt
@@ -18,14 +18,11 @@
 package com.example.tidy.viewModels
 
 import androidx.lifecycle.ViewModel
-import androidx.navigation.NavController
 import com.example.tidy.DbOperation
 import com.example.tidy.Task
-import com.example.tidy.constants.Routes
 
 class AddTaskScreenViewModel(
     private val dbOperation: DbOperation,
-    private val navController: NavController,
 ) : ViewModel() {
     private var childFlag: Boolean = false
     var parentTaskId: Long = 0
@@ -53,9 +50,5 @@ class AddTaskScreenViewModel(
             dbOperation.addChild(childId = i, parentId = parentTaskId)
             childFlag = false
         }
-    }
-
-    fun editTask(task: Task) {
-        navController.navigate("${Routes.ADD_TASK}/${task.id}")
     }
 }
